+++ tests/fixtures/source/IncompleteDataConstructor.xs
@@ -1,0 +1,16 @@
+// SPDX-FileCopyrightText: 2026 Leitwolf <xs-lang.chess031@slmails.com>
+// SPDX-License-Identifier: Apache-2.0
+
+data Point {
+    x: Long;
+    y: Long;
+
+    Point(x: Long) {
+        self.x = x;
+    }
+}
+
+fn main() -> Long {
+    point: Point = Point(1);
+    return point.x;
+}