+++ xslang/sources/lib.rs
@@ -5,9 +5,9 @@
 
 pub mod codegen;
 pub mod compiler_core;
-pub mod gc;
 pub mod hir;
 pub mod mir;
 pub mod mono;
 pub(crate) mod text;
+pub mod xgc;
 pub mod xlil;